fix: clear all non-sign-conversion compiler warnings
Eliminate every warning except the long-tail -Wsign-conversion (471 on
clang, 422 on gcc). Verified: 0 non-sign-conversion warnings under both
the faithful make-metal clang flags and gcc, for all C TUs. Real build
links cleanly.

model.c:
- Remove 5 vestigial debug-stat blocks (qsum/qmax, lsum/lmax/lmin, qsum,
  dsum/dmax, xsum/xmax) that computed per-layer statistics and discarded
  them -- leftovers from removed fprintf debug prints, guarded only by
  l==0/pos==0 conditions.
- Remove vestigial gbytes/budget computation in model_unlock_layers
  (copy-paste from model_lock_layers, where it is genuinely used; unlock
  works purely off the re-computed ranges).
- Cast the dequantize_row() count arg to int explicitly (3 sites):
  laurel_l/laurel_r/router dequant used a size_t product for an int param.

quant.c:
- Add explicit casts for 18 narrowing conversions across the Q8_0/Q8_K/
  Q4_0/Q3_K/Q1_0 quantize + vec_dot paths (int32_t->int8_t, int->int16_t,
  int->uint8_t nibble packing, int->float accumulator scales). All
  preserve exact prior semantics -- the values are already in range.
- Guard the get_scale_shuffle_k6[144] lookup table with the same
  #if defined(PICOLM_AVX2) as its (sole) consumers, so it is no longer
  flagged -Wunused-const-variable on non-AVX2 builds (ARM, SSE, etc.).

server.c:
- Cast float sampler params (temperature/top_p/repeat_penalty) to double
  at the cJSON_AddNumberToObject() calls; they stay float for sampler_init.

tensor.h:
- Add missing trailing newline (-Wnewline-eof).

+++ picolm/server.c
@@ -2419,10 +2419,10 @@ static void handle_llama_completion(SOCKET sock, const char *request_body) {
     /* Build generation_settings JSON */
     cJSON *gen_settings = cJSON_CreateObject();
     cJSON_AddNumberToObject(gen_settings, "n_predict", n_predict);
-    cJSON_AddNumberToObject(gen_settings, "temperature", temperature);
-    cJSON_AddNumberToObject(gen_settings, "top_p", top_p);
+    cJSON_AddNumberToObject(gen_settings, "temperature", (double)temperature);
+    cJSON_AddNumberToObject(gen_settings, "top_p", (double)top_p);
     cJSON_AddNumberToObject(gen_settings, "top_k", top_k);
-    cJSON_AddNumberToObject(gen_settings, "repeat_penalty", repeat_penalty);
+    cJSON_AddNumberToObject(gen_settings, "repeat_penalty", (double)repeat_penalty);
     cJSON_AddNumberToObject(gen_settings, "repeat_last_n", repeat_last_n);
     cJSON_AddNumberToObject(gen_settings, "seed", seed);
